i-oic-confluent/requirements.txt
confluent-kafka==2.14.0
fastapi==0.136.1
uvicorn==0.46.0
requests==2.33.1
python-dotenv==1.2.2
fastmcp>=0.3.0